backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0

# 유정우 제발 정신차려 vscode 아래에 UTF-8로 바꿔야 적용됨 
django-cors-headers==4.0.0

psycopg2-binary==2.9.6
redis==4.5.5
celery==5.2.7
flower==1.2.0
requests==2.30.0
python-dotenv==1.0.0

# 데이터 분석/딥러닝
numpy>=1.26.0
pandas==2.0.3
scikit-learn>=1.7.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
pillow==10.0.0
xgboost>=1.7.0
opencv-python==4.8.0.76
lightgbm==4.6.0
# torch==2.0.1
# torchvision==0.15.2
# tensorflow==2.13.0
# xgboost==1.7.6
# lightgbm==4.0.0
# catboost==1.2

# ✅ 모니터링용 추가
django-prometheus==2.2.0
slack-sdk==3.21.3

django-celery-beat

pydicom

# 유정우가 추가
djangorestframework-simplejwt 